:art: Make dashboard columns resizable
